third_party/WebKit/LayoutTests/http/tests/devtools/sources/pretty-print-javascript-9.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/sources-test.js"></script>

<script>

function test() {
  var testJSFormatter = SourcesTestRunner.testPrettyPrint.bind(InspectorTest, 'text/javascript');

  TestRunner.runTestSuite([
    function parenthesizedExpressions(next) {
      var mappingQueries = ['if', '((a))', '((b));', 'else', '(c)'];
      testJSFormatter('if((a))((b));else (c);', mappingQueries, next);
    },

    function objectDesctructuring(next) {
      var mappingQueries = ['let', 'y', 'getXYFromTouchOrPointer', 'e'];
      testJSFormatter('let{x,y}=getXYFromTouchOrPointer(e);', mappingQueries, next);
    },

    function objectDesctructuringInFunctionExpression(next) {
      var mappingQueries = ['test', 'function', 'foo'];
      testJSFormatter('var test = function({x,y}){foo(x,y);}', mappingQueries, next);
    },
  ]);
}

</script>

</head>

<body onload="runTest()">
<p>Verifies JavaScript pretty-printing functionality.</p>
</body>
</html>
