third_party/WebKit/LayoutTests/http/tests/devtools/sources/pretty-print-javascript-9-expected.txt

Verifies JavaScript pretty-printing functionality.


Running: parenthesizedExpressions
====== 8< ------
if ((a))
    ((b));
else
    (c);

------ >8 ======
Correct mapping for <if>
Correct mapping for <((a))>
Correct mapping for <((b));>
Correct mapping for <else>
Correct mapping for <(c)>

Running: objectDesctructuring
====== 8< ------
let {x, y} = getXYFromTouchOrPointer(e);

------ >8 ======
Correct mapping for <let>
Correct mapping for <y>
Correct mapping for <getXYFromTouchOrPointer>
Correct mapping for <e>

Running: objectDesctructuringInFunctionExpression
====== 8< ------
var test = function({x, y}) {
    foo(x, y);
}

------ >8 ======
Correct mapping for <test>
Correct mapping for <function>
Correct mapping for <foo>
